backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Marker PDF
marker-pdf>=0.2.0

# LLM API
google-generativeai>=0.3.0  # For Google Gemini API

# other
pydantic>=2.0.0
python-dotenv>=1.0.0
